+++ org.restlet.java/org.restlet.ext.gae/src/main/java/org/restlet/ext/gae/GaeEnroler.java
@@ -28,14 +28,6 @@ public class GaeEnroler implements Enroler {
     private Role adminRole;
 
     /**
-     * Default constructor. It defines an administrator role, which name is
-     * "admin".
-     */
-    public GaeEnroler() {
-        this("admin", "Administrator of the current application.");
-    }
-
-    /**
      * Constructor.
      * 
      * @param application
@@ -71,32 +63,6 @@ public class GaeEnroler implements Enroler {
      */
     public GaeEnroler(Role adminRole) {
         setAdminRole(adminRole);
-    }
-
-    /**
-     * Constructor.
-     * 
-     * @param adminRoleName
-     *            The name of the administrator role.
-     * @deprecated Use {@link #GaeEnroler(Application, String)} instead.
-     */
-    @Deprecated
-    public GaeEnroler(String adminRoleName) {
-        this(adminRoleName, "Administrator of the current application.");
-    }
-
-    /**
-     * Constructor.
-     * 
-     * @param adminRoleName
-     *            The name of the administrator role.
-     * @param adminRoleDescription
-     *            The description of the administrator role.
-     * @deprecated Use {@link #GaeEnroler(Application, String, String)} instead.
-     */
-    @Deprecated
-    public GaeEnroler(String adminRoleName, String adminRoleDescription) {
-        this(new Role(adminRoleName, adminRoleDescription));
     }
 
     /**