+++ org.restlet.java/org.restlet/src/main/java/org/restlet/service/LogService.java
@@ -290,17 +290,6 @@ public class LogService extends Service {
 	}
 
 	/**
-	 * Indicates if the debugging mode is enabled. False by default.
-	 * 
-	 * @return True if the debugging mode is enabled.
-	 * @deprecated Rely on {@link Application#isDebugging()} instead.
-	 */
-	@Deprecated
-	protected boolean isDebugging() {
-		return debugging;
-	}
-
-	/**
 	 * Indicates if the identity check (as specified by RFC1413) is enabled. Default
 	 * value is false.
 	 * 
@@ -321,17 +310,6 @@ public class LogService extends Service {
 	public boolean isLoggable(Request request) {
 		return (getLoggableTemplate() == null) ? true
 				: getLoggableTemplate().match(request.getResourceRef().getTargetRef().toString()) > 0;
-	}
-
-	/**
-	 * Indicates if the debugging mode is enabled.
-	 * 
-	 * @param debugging True if the debugging mode is enabled.
-	 * @deprecated Rely on {@link Application#setDebugging(boolean)} instead.
-	 */
-	@Deprecated
-	protected void setDebugging(boolean debugging) {
-		this.debugging = debugging;
 	}
 
 	/**